requirements.txt
selenium==4.1.3
webdriver-manager==3.5.4
python-dotenv==0.20.0
requests==2.27.1
sqlitedict==2.0.0
